Moved components in seperate folders
