+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import { Header } from './components/Header';
-import { DarkModeContainer } from './components/DarkModeContainer';
+import { DarkModeContainer } from './components/utils/DarkModeContainer';
 import { UpcomingEvents } from './components/UpcomingEvents';
 import { LatestMemes } from './components/LatestMemes';
 import { LatestBlasts } from './components/LatestBlasts';
@@ -11,6 +11,7 @@ const MS_PER_COMPONENT = SECONDS_PER_COMPONENT * 1000;  // Convert seconds to mi
 function App() {
   // array of main components to cycle through
   const components = [
+    <UpcomingEvents key={0} />,
     <>
       <div className='flex justify-between p-3 mb-5 text-4xl font-bold bg-white border dark:border-b-gray-700 dark:bg-[#111827] dark:border-0 dark:border-b-[1px] dark:text-white border-b-light-grey px-28'>
         <div>#memeogvinogklinoggrin2</div>
@@ -21,10 +22,9 @@ function App() {
         <div className='w-1 -mt-10 border-l -z-10 border-light-grey'></div>
         <LatestBlasts />
       </div>
-    </>,
-    <UpcomingEvents key={0} />,
+    </>
   ];
-  const [currentComponentIndex, setCurrentComponentIndex] = useState(0);
+  const [currentComponentIndex, setCurrentComponentIndex] = useState(1);
   const [opacity, setOpacity] = useState(1);
   const [millisecondsLeft, setMillisecondsLeft] = useState(MS_PER_COMPONENT);
