+++ src/components/cards/MemeCard.tsx
@@ -1,12 +1,10 @@
-import { formatSlackDate } from "../lib/date";
-import { MemeType } from "../lib/types";
+import { formatSlackDate } from "../../lib/date";
+import { MemeType } from "../../lib/types";
+import { BaseCard } from "./BaseCard";
 
 export const MemeCard = ({ meme, width }: { meme: MemeType, width: number }) => {
   return (
-    <div
-      className="relative flex flex-col justify-center overflow-hidden bg-white border border-gray-200 shadow rounded-xl dark:bg-gray-800 dark:border-gray-700 -z-10"
-      key={meme.url}
-    >
+    <BaseCard>
       <div className="flex items-center w-full gap-4 px-4 py-3">
         <img
           className="w-10 h-10 rounded-full"
@@ -26,6 +24,6 @@ export const MemeCard = ({ meme, width }: { meme: MemeType, width: number }) => 
         alt={`Meme ${meme.url}`}
         style={{ width: `${width}px` }}
       />
-    </div>
+    </BaseCard>
   )
 }
